Added a preference pane
I created a MDPreferenceController class which which takes care of instantiating the Preference.xib file and makes the preference view accessible to other classes by uses of the Singleton (anti?)pattern.

In MDMissingDrawer i added a method that swaps instance methods of OakPreferencesManager to my implementation of them in order to add the Missing Drawer preference tab to the preference window.

The preference view using NSUserDefault's to set and read the values of the background color when active and inactive.

I changed NSWindowController+MDAdditions.m so it sets the background according to the value stored in the NSUserDefault.

Removed the existing code that tried to set the background.